Can now store larger datasets
Previously, the script made several manipulations of the very large datasets that were beign saved for the .dat file. In order to do this Matlab first has to make a copy of the data, and cannot do "in-position" manipulations or deletions. Now all the memory is stored from the beginning in the format it will be saved at so no copies need to be made. Pre-Process is also no longer calling "writedat" function because it also has several manipulations that require array manipulation.